endoviz-main/pyproject.toml
[project]
name = "endoviz"
version = "0.1.0"
description = "Node embedding and visualization pipeline for Reactome pathway data"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "some ppl"}
]
keywords = ["bioinformatics", "embeddings", "network-analysis", "reactome", "visualization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Core dependencies (required for term-based pipeline)
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.26.0,<2.2",  # <2.2 required for UMAP compatibility
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "umap-learn>=0.5.0",
    "pacmap>=0.7.0",
    "pyarrow>=10.0.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
]

# Optional dependencies for graph-based pipeline
[project.optional-dependencies]
graph = [
    "networkx>=3.0",
    "node2vec>=0.5.0",
]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

